+++ Tests/FeatureComposerTests/DomainTests/InteractorTests/CounterInteractors/CounterInteractor.swift
@@ -13,7 +13,7 @@ struct CounterInteractor: Interactor {
     }
     
     var body: some InteractorOf<Self> {
-        Interact<State, Action>(defaultValue: State(count: 0)) { state, action in
+        Interact<State, Action>(initialValue: State(count: 0)) { state, action in
             switch action {
             case .increment:
                 state.count += 1
